+++ src/drivers/SimpleKeyValueDriver.ts
@@ -1,5 +1,6 @@
 import { SyncFeatureManagerDriver } from '../base-drivers/SyncFeatureManagerDriver'
 import { CommonValueParams, ValueReturnType } from '../types/common.types'
+import { deriveValue } from '../utils'
 
 /**
  * Takes in a key / value mapping as the configuration. Supports sync and async operations.
@@ -26,10 +27,9 @@ export class SimpleKeyValueDriver<
     K extends string & keyof Flags,
     Params extends CommonValueParams<Flags, K> | undefined = undefined,
   >(key: K, params?: Params): ValueReturnType<Flags, K, Params> {
-    return (
-      (this.conf[key] as Flags[K]) ??
-      params?.defaultValue ??
-      (null as ValueReturnType<Flags, K, Params>)
+    return deriveValue<Flags, K, Params>(
+      this.conf[key] as Flags[K],
+      params?.defaultValue
     )
   }